prov/efa: cleanup rxr_flags, remove queuing from rx/tx state
Cleanup rxr_flags as some of those flags are no longer used and these
flags are not sent over the wire.

Reuse those old flags to move queued and rnr state out of the state enum
and instead set a bit to mark that this entry is queued. This makes the
logic clearer and more general by avoiding the need to transition the
state back after the entry is no longer queued. That wasn't being done
for all of the various uses of tx/rx entries.

One thing we do have to be careful about in the future is to not attempt
to progress entries that are in this state. Most of the protocol will
only try to post one control message at a time, and then we implicitly
block on that packet being sent. In the long segmented message case we
can check for this flag before sending more packets.

Signed-off-by: Robert Wespetal <wesper@amazon.com>